backend/miproyecto_django/requirements.txt
Django>=4.2.0
python-decouple>=3.8  # Para manejar variables de entorno
Pillow>=10.0.0  # Para manejo de imágenes
django-crispy-forms>=2.0  # Para formularios más atractivos
crispy-bootstrap5>=0.7  # Template pack para crispy-forms
django-allauth>=0.57.0  # Para autenticación
whitenoise>=6.5.0  # Para servir archivos estáticos
gunicorn>=21.2.0  # Servidor WSGI para producción
mysqlclient>=2.2.0  # Para MySQL
django-debug-toolbar>=4.2.0  # Para desarrollo
django-sass-processor>=1.1  # Para procesar archivos SASS/SCSS
libsass>=0.22.0  # Dependencia requerida por django-sass-processor
django-localflavor>=3.0  # Para validación de campos específicos por país
django-countries>=7.5.1  # Para manejo de países
django-encrypted-model-fields>=0.6.3  # Para campos encriptados
python-dateutil>=2.8.2  # Para manejo de fechas
django-widget-tweaks>=1.5.0  # Para personalización de widgets de formularios 
python-dotenv
django_ratelimit
djangorestframework
django-cors-headers
djangorestframework-simplejwt
pandas>=2.0.0
openpyxl>=3.1.2
unidecode>=1.3
pytest
django-filter